requirements.txt
# Framework principal
Flask==3.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
python-dotenv==1.0.0
Flask-Cors==4.0.0
Flask-Static-Digest==0.2.1
whitenoise==6.5.0

# Dependencias de Flask
MarkupSafe==2.1.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

# Bases de datos
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.20

# MongoDB
pymongo==4.6.1
Flask-PyMongo==2.3.0
dnspython==2.3.0

# PostgreSQL
psycopg2-binary==2.9.9

# Servicios externos
stripe==7.6.0

# Generación de documentos
reportlab==4.0.7
